Moved, renamed, and clean up
Rename game to Game.java
Renamed Player to Player.java
Moved Game.java and Player.java into the firstpackagejpp directory
Added imports
Cleaned up types
