--- Jones_Parham_Porter/src/jones_parham_porter/Jones_Parham_Porter.java
@@ -1,28 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-
-/* 
-This is a test to see if this works. Lets hope.
-hello world testing 123
-I got your message.  This is Kyle.
-*/
-package jones_parham_porter;
-
-/**
- *
- * @author Kyle Jones
- */
-public class Jones_Parham_Porter {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        // TODO code application logic here
-    }
-    
-}
