+++ oregontrail1/src/firstpackagejpp/Game.java
@@ -1,0 +1,53 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package firstpackagejpp;
+
+import java.io.Serializable;
+
+/**
+ *
+ * @author parhe
+ */
+public class Game implements Serializable {
+    
+    private double totalTime;
+    private int noPeople;
+    private Player player; 
+
+    public Game() {
+    }
+
+    public double getTotalTime() {
+        return totalTime;
+    }
+
+    public void setTotalTime(double totalTime) {
+        this.totalTime = totalTime;
+    }
+
+    public int getNoPeople() {
+        return noPeople;
+    }
+
+    public void setNoPeople(int noPeople) {
+        this.noPeople = noPeople;
+    }
+
+    public Player getPlayer() {
+        return player;
+    }
+
+    public void setPlayer(Player player) {
+        this.player = player;
+    }
+
+    
+
+    
+    
+    
+    
+}
